+++ changelog.md
@@ -3,3 +3,6 @@
 - Added basic login functionality
 - Add explanation for credential requirement
 - Added ability to login to Steam
+- Added personal settings encryption
+- Added base api module structure
+- Added basic match window